# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Technical Test JARIVIS - IQBAL IKHLASUL AMAL",
        "contact": {
            "name": "Iqbal Ikhlasul Amal",
            "email": "iqbaliamal@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/kompetensi-dasar-point/list": {
            "get": {
                "tags": [
                    "Kompetensi Dasar Point"
                ],
                "summary": "Retrieve all kompetensi-dasar-points",
                "operationId": "da1ab4589094fb4e14a07fdcda99b807",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kompetensi-dasar-point/store": {
            "post": {
                "tags": [
                    "Kompetensi Dasar Point"
                ],
                "summary": "Create a new kompetensi-dasar-point",
                "operationId": "8b916486ce2aee8200d73d2ebe3a17ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kompetensi_dasar_detail_id",
                                    "name"
                                ],
                                "properties": {
                                    "kompetensi_dasar_detail_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "kompetensi_dasar_detail_id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        },
                                        "deleted_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kompetensi-dasar-point/detail/{id}": {
            "get": {
                "tags": [
                    "Kompetensi Dasar Point"
                ],
                "summary": "Retrieve a kompetensi-dasar-point",
                "operationId": "08a9daa07105669a3dc10eab722d7e27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "kompetensi-dasar-point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kompetensi-dasar-point/update/{id}": {
            "put": {
                "tags": [
                    "Kompetensi Dasar Point"
                ],
                "summary": "Update a kompetensi-dasar-point",
                "operationId": "53b14ce6ddbb9544f2b5aeaee17cdefb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "kompetensi-dasar-point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kompetensi_dasar_detail_id",
                                    "name"
                                ],
                                "properties": {
                                    "kompetensi_dasar_detail_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "kompetensi_dasar_detail_id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        },
                                        "deleted_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kompetensi-dasar-point/delete/{id}": {
            "delete": {
                "tags": [
                    "Kompetensi Dasar Point"
                ],
                "summary": "Delete a kompetensi-dasar-point",
                "operationId": "3f0a1d11d5065801a41c1478d3a57c9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "kompetensi-dasar-point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}